feat: Update product navigation in GastroEntorologiaPage
Refactor the product navigation in the GastroEntorologiaPage to fix the routing issue. The router.push('/home/gastro/gastroentorologia/products') line was commented out, causing the button click to not navigate to the correct page. This commit fixes the issue by uncommenting the router.push line.

style: Adjust layout on GastroEntorologiaPage

Center align the button container and adjust the padding for improved layout on the GastroEntorologiaPage. This change enhances the visual appearance and user experience of the page.

feat: Add Gastroenterology specialty image

Add the Gastroenterology specialty image to the assets folder. The image is named "green.png" and is located in the src/assets/img/general/desktop/png directory. This image will be used to enhance the visual representation of the Gastroenterology specialty.

Refactor: Update GastroContainer and ImageCarousel

Update the GastroContainer and ImageCarousel components to support the initial slide index and format slides from all products. This refactor improves the functionality and flexibility of the components, allowing for better customization and usability.